requirements.txt
# packages to run the app for both environments
mitmproxy>=10.1.5
watchdog>=3.0.0
# for downloading the player and libmpv
feedparser>=6.0.10
py7zr>=0.20.6
py-cpuinfo>=9.0.0
requests>=2.31.0
# for EPG
lxml>=4.9.4
